+++ duality/__init__.py
@@ -1,35 +1,38 @@
 '''
 
-This is a connection to the main folder of the duality package.
+This is a connection to the main folder of the duality library.
 
-AVAILABLE MODULES IN THE PACKAGE:
+AVAILABLE MODULES IN THE LIBRARY:
 
-	HUB
-	---
-	duality.hub is a hub of data manipulation tools.
-		print(help(duality.hub)) in order to see available features.
+	TOOLKIT
+	-------
+
+	duality.toolkit is a set of data manipulation tools that can be directly accessed from the main module.
+		print(help(duality.toolkit)) in order to see available features.
 
 	MONTECARLO
-	---
+	----------
+
 	duality.MonteCarlo is a module for performing the Monte Carlo simulation over the defined data with a lot of useful features.
 		print(help(duality.MonteCarlo)) in order to see available features.
 
 	EOQ
 	---
+	
 	duality.EOQ is a module for finding an Economic order quantity over the defined data with a lot of useful features.
 		print(help(duality.EOQ)) in order to see available features.
 
 '''
 
-#ignore __pycache__ from forming inside the package directory.
+#ignore __pycache__ from forming inside the library directory.
 import sys
 sys.dont_write_bytecode = True
 
-#imports essential functions from the duality package.
-from duality.eoq.eoq import EOQ
-from duality.hub import hub
-from duality.hub.hub import *
-from duality.montecarlo.montecarlo import MonteCarlo
+#imports essential functions from the duality library.
+from duality.hub import toolkit
+from duality.hub.toolkit import *
+from duality.objects.eoq import EOQ
+from duality.objects.montecarlo import MonteCarlo
 from duality.misc._meta import (
 	__author__,
 	__copyright__,
@@ -40,3 +43,10 @@ from duality.misc._meta import (
 	__contact__,
 	__donate__
 )
+
+#all relevant contents.
+__all__ = [
+	toolkit,
+	MonteCarlo,
+	EOQ,
+]