+++ duality/objects/__init__.py
@@ -1,0 +1,9 @@
+#import all relevant contents from the associated module.
+from duality.objects.montecarlo import MonteCarlo
+from duality.objects.eoq import EOQ
+
+#all relevant contents.
+__all__ = [
+    MonteCarlo,
+    EOQ,
+]